+++ packages/solana-wallet-snap/src/core/services/subscriptions/SignatureMonitor.ts
@@ -1,4 +1,4 @@
-import type { Logger } from '@metamask/snap-networks-utils';
+import type { AnalyticsService, Logger } from '@metamask/snap-networks-utils';
 import { assert, string } from '@metamask/superstruct';
 import { Duration } from '@metamask/utils';
 import { signature as asSignature } from '@solana/kit';
@@ -14,8 +14,6 @@ import type {
 import type { Network } from '../../constants/solana';
 import { trackError } from '../../utils/errors';
 import type { AccountsService } from '../accounts/AccountsService';
-import type { AnalyticsService } from '../analytics/AnalyticsService';
-import type { ConfigProvider } from '../config';
 import { SUPPORTED_NETWORKS } from '../config/ConfigProvider';
 import type { SolanaConnection } from '../connection';
 import type { TransactionsService } from '../transactions';
@@ -31,8 +29,6 @@ export class SignatureMonitor {
   readonly #analyticsService: AnalyticsService;
 
   readonly #connection: SolanaConnection;
-
-  readonly #configProvider: ConfigProvider;
 
   readonly #logger: Logger;
 
@@ -52,7 +48,6 @@ export class SignatureMonitor {
     transactionsService: TransactionsService,
     analyticsService: AnalyticsService,
     connection: SolanaConnection,
-    configProvider: ConfigProvider,
     logger: Logger,
   ) {
     this.#subscriptionService = subscriptionService;
@@ -60,7 +55,6 @@ export class SignatureMonitor {
     this.#transactionsService = transactionsService;
     this.#analyticsService = analyticsService;
     this.#connection = connection;
-    this.#configProvider = configProvider;
     this.#logger = logger.withPrefix('[✍️ SignatureMonitor]');
 
     this.#bindHandlers();
@@ -186,14 +180,13 @@ export class SignatureMonitor {
       switch (commitment) {
         case 'confirmed':
         case 'finalized':
-          await this.#analyticsService.trackEventTransactionFinalized(
-            account,
-            transaction,
-            {
-              scope: network,
-              origin,
-            },
-          );
+          await this.#analyticsService.trackTransactionFinalized({
+            origin,
+            accountType: account.type,
+            chainIdCaip: transaction.chain,
+            transactionStatus: transaction.status,
+            transactionType: transaction.type,
+          });
           break;
         default:
           this.#logger.warn(`⚠️ Commitment ${commitment} not supported`);