+++ packages/stellar-wallet-snap/src/context.ts
@@ -1,4 +1,8 @@
-import { State } from '@metamask/snap-networks-utils';
+import {
+  AnalyticsService,
+  State,
+  InMemoryCache,
+} from '@metamask/snap-networks-utils';
 import { assert, object } from '@metamask/superstruct';
 
 import { AppConfig } from './config';
@@ -16,6 +20,7 @@ import { OnAddressInputHandler } from './handlers/clientRequest/onAddressInput';
 import { OnAmountInputHandler } from './handlers/clientRequest/onAmountInput';
 import { SignAndSendTransactionHandler } from './handlers/clientRequest/signAndSendTransaction';
 import { SignProofOfOwnershipHandler } from './handlers/clientRequest/signProofOfOwnership';
+import { SignProofOfOwnershipBatchHandler } from './handlers/clientRequest/signProofOfOwnershipBatch';
 import type { ICronjobRequestHandler } from './handlers/cronjob/api';
 import { BackgroundEventMethod } from './handlers/cronjob/api';
 import {
@@ -39,7 +44,6 @@ import {
   AssetMetadataRepository,
   AssetMetadataService,
 } from './services/asset-metadata';
-import { InMemoryCache } from './services/cache';
 import { NetworkService } from './services/network';
 import {
   OnChainAccountRepository,
@@ -59,7 +63,7 @@ import {
 } from './services/transaction-scan';
 import { WalletService } from './services/wallet';
 import { ConfirmationUXController } from './ui/confirmation/controller';
-import { logger, noOpLogger } from './utils';
+import { getSnapProvider, logger, noOpLogger, trackError } from './utils';
 
 assert(AppConfig, object());
 
@@ -73,6 +77,12 @@ const transactionRepository = new TransactionRepository(state);
 const assetMetadataRepository = new AssetMetadataRepository(state);
 
 /** ------------------------------ Services  ------------------------------ */
+const analyticsService = new AnalyticsService({
+  getSnapProvider,
+  logger,
+  trackError,
+});
+
 const appCache = new InMemoryCache(noOpLogger);
 const networkService = new NetworkService({ logger, cache: appCache });
 
@@ -213,6 +223,7 @@ const trackTransactionHandler = new TrackTransactionHandler({
   networkService,
   synchronizeService,
   accountService,
+  analyticsService,
 });
 
 const syncAccountsHandler = new SyncAccountsHandler({
@@ -248,6 +259,7 @@ const changeTrustOptHandler = new ChangeTrustOptHandler({
   assetMetadataService,
   transactionService,
   confirmationUIController,
+  analyticsService,
 });
 
 const onAddressInputHandler = new OnAddressInputHandler();
@@ -264,6 +276,7 @@ const signAndSendTransactionHandler = new SignAndSendTransactionHandler({
   accountResolver,
   transactionService,
   assetMetadataService,
+  analyticsService,
 });
 
 const confirmSendHandler = new ConfirmSendHandler({
@@ -272,6 +285,7 @@ const confirmSendHandler = new ConfirmSendHandler({
   transactionService,
   assetMetadataService,
   confirmationUIController,
+  analyticsService,
 });
 
 const computeFeeHandler = new ComputeFeeHandler({
@@ -285,6 +299,12 @@ const signProofOfOwnershipHandler = new SignProofOfOwnershipHandler({
   accountResolver,
 });
 
+const signProofOfOwnershipBatchHandler = new SignProofOfOwnershipBatchHandler({
+  logger,
+  accountService,
+  walletService,
+});
+
 const clientRequestMethodHandlers: Record<
   ClientRequestMethod,
   IClientRequestHandler
@@ -296,6 +316,8 @@ const clientRequestMethodHandlers: Record<
   [ClientRequestMethod.SignAndSendTransaction]: signAndSendTransactionHandler,
   [ClientRequestMethod.ComputeFee]: computeFeeHandler,
   [ClientRequestMethod.SignProofOfOwnership]: signProofOfOwnershipHandler,
+  [ClientRequestMethod.SignProofOfOwnershipBatch]:
+    signProofOfOwnershipBatchHandler,
 };
 
 const clientRequestHandler = new ClientRequestHandler({