+++ packages/tron-wallet-snap/jest.config.js
@@ -1,0 +1,28 @@
+const baseConfig = require('../../jest.config.packages');
+
+module.exports = {
+  ...baseConfig,
+  preset: '@metamask/snaps-jest',
+  transform: {
+    '^.+\\.(t|j)sx?$': 'ts-jest',
+  },
+  moduleNameMapper: {
+    ...baseConfig.moduleNameMapper,
+    '\\.svg$': 'jest-transform-stub',
+  },
+  testMatch: ['**/src/**/?(*.)+(spec|test).[tj]s?(x)'],
+  setupFilesAfterEnv: [
+    ...baseConfig.setupFilesAfterEnv,
+    '<rootDir>/jest.setup.ts',
+  ],
+
+  // An object that configures minimum threshold enforcement for coverage results
+  coverageThreshold: {
+    global: {
+      branches: 69.96,
+      functions: 75.86,
+      lines: 82.61,
+      statements: 82.62,
+    },
+  },
+};